Add files via upload
This Matlab script is for my Youtube Signal Processing series entitled,
“A Journey Into Music and Sound”.

Part 1 – Intro
Part 2 – Matlab IDE and wavread
Part 3 – Playing Music in Reverse and Increasing Volume
Part 4 – Speeding Up and Slowing Down Music
This file opens a music file and allows you to analyze the music file in the time and frequency domains. In addition you can increase the amplitude and frequencies of the music opened.